fix codecheck points
