+++ .env
@@ -1,0 +1,1 @@
+AWS_PROFILE=harish-exasol-platform
