+++ tests/tests/e2e/test_cli_smoke.py
@@ -6,25 +6,6 @@ import pytest
 pytestmark = [pytest.mark.e2e]
 
 
-def test_unknown_flag_exits_nonzero_with_usage(exasol_path: str) -> None:
-    """Unsupported flag must exit non-zero with usage and no stack trace."""
-    # Given the launcher binary
-    # When called with a flag that does not exist
-    proc = subprocess.run(
-        [exasol_path, "--definitely-not-a-flag"],
-        capture_output=True,
-        text=True,
-        check=False,
-    )
-
-    # Then it exits non-zero and surfaces usage without a stack trace
-    assert proc.returncode != 0
-    combined = (proc.stdout + proc.stderr).lower()
-    assert "unknown flag" in combined or "usage" in combined
-    assert "traceback" not in combined
-    assert "panic:" not in combined
-
-
 @pytest.mark.skipif(
     not sys.platform.startswith("win"), reason="Windows-only shell-equivalence test"
 )
