+++ tests/tests/integration/test_cli.py
@@ -1,9 +1,29 @@
 # Copyright 2026 Exasol AG
 # SPDX-License-Identifier: MIT
 
+import subprocess
 from pathlib import Path
 
 from .helpers import first_infrastructure_preset_id_or_skip, run_command
+
+
+def test_unknown_flag_exits_nonzero_with_usage(exasol_path: str) -> None:
+    """Unsupported flag must exit non-zero with usage and no stack trace."""
+    # Given the launcher binary
+    # When called with a flag that does not exist
+    proc = subprocess.run(
+        [exasol_path, "--definitely-not-a-flag"],
+        capture_output=True,
+        text=True,
+        check=False,
+    )
+
+    # Then it exits non-zero and surfaces usage without a stack trace
+    assert proc.returncode != 0
+    combined = (proc.stdout + proc.stderr).lower()
+    assert "unknown flag" in combined or "usage" in combined
+    assert "traceback" not in combined
+    assert "panic:" not in combined
 
 
 def test_help_flag(exasol_path: str) -> None:
